backend/requirements.txt
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0

uwsgi==2.0.17
python-docx==0.8.10
pdfminer.six==20181108
psycopg2==2.8
whitenoise==3.3.1
gunicorn==19.9.0
psycopg2-binary==2.8.4

Pillow==7.0.0